+++ packages/go_router/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 17.2.0
 
-- Adds `overrideOnExit` parameter to `TypedGoRoute` and `TypedRelativeGoRoute` to control whether the `onExit` method of route data classes should be invoked.
+- Adds `hasOverriddenOnExit` parameter to `TypedGoRoute` and `TypedRelativeGoRoute` to enable custom exit handling in type-safe routes. When set to `true`, the overridden `onExit` method in route data classes (extending `GoRouteData` or `RelativeGoRouteData`) will be invoked when the route is removed from GoRouter's route history.
 
 ## 17.1.0